# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sprinklerGo API",
    "description": "REST-API der sprinklerGo-Bewässerungssteuerung (Go-Port von sprinklers_pi). Zonen- und Programm-IDs sind 0-basierte Indizes; Startzeiten sind Minuten seit lokaler Mitternacht; Laufzeiten sind Minuten (0-255).",
    "version": "1.0.0",
    "license": { "name": "GPL-2.0", "identifier": "GPL-2.0-only" }
  },
  "paths": {
...
    "/api/zones/{id}": {
      "put": {
        "summary": "Zone ändern",
        "operationId": "updateZone",
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ZoneInput" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/zones/{id}/manual": {
      "post": {
        "summary": "Zone manuell an/aus",
        "operationId": "manualZone",
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["on"],
                "properties": {
                  "on": { "type": "boolean" },
                  "minutes": {
                    "type": "integer", "minimum": 0, "maximum": 1440,
                    "description": "Laufzeitbegrenzung; weggelassen = Einstellung manualTimerMinutes, 0 = unbegrenzt"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/schedules": {
      "get": {
        "summary": "Programmliste inkl. nächster Läufe",
        "operationId": "getSchedules",
        "responses": {
          "200": {
            "description": "Alle Programme",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schedules": { "type": "array", "items": { "$ref": "#/components/schemas/Schedule" } }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Programm anlegen",
        "operationId": "createSchedule",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleInput" } } }
        },
        "responses": {
          "201": {
            "description": "Angelegt",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "id": { "type": "integer" } } }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/schedules/{id}": {
      "get": {
        "summary": "Einzelnes Programm",
        "operationId": "getSchedule",
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": {
            "description": "Programm",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Schedule" } } }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "put": {
        "summary": "Programm ändern",
        "operationId": "updateSchedule",
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleInput" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "summary": "Programm löschen (nachfolgende IDs rücken auf)",
        "operationId": "deleteSchedule",
        "parameters": [{ "$ref": "#/components/parameters/id" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/quickrun": {
      "post": {
        "summary": "Schnellstart: bestehendes Programm sofort ODER Ad-hoc-Laufzeiten",
        "description": "Genau eines der Felder scheduleId/durations angeben. Ein Programmstart wendet saisonale und Wetter-Anpassung an; Ad-hoc-Laufzeiten laufen unangepasst.",
        "operationId": "quickRun",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scheduleId": { "type": "integer" },
                  "durations": { "type": "array", "items": { "type": "integer", "minimum": 0, "maximum": 255 } }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/rain-delay": {
      "put": {
        "summary": "Regenpause setzen oder aufheben",
        "description": "Unterdrückt Programmstarts für die angegebene Dauer; manuelle Läufe und Schnellstarts sind nicht betroffen. hours=0 hebt die Pause auf. Überlebt einen Neustart.",
        "operationId": "setRainDelay",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["hours"],
                "properties": { "hours": { "type": "integer", "minimum": 0, "maximum": 336 } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Gesetzt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "rainDelayUntil": { "type": "integer", "description": "Unix-Sekunden, 0 = keine Pause" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/auth": {
      "get": {
        "summary": "Auth-Status (immer erreichbar)",
        "operationId": "getAuth",
        "responses": {
          "200": {
            "description": "Status; Token-Liste nur für angemeldete Aufrufer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": { "type": "boolean" },
                    "loggedIn": { "type": "boolean" },
                    "hasPassword": { "type": "boolean" },
                    "tokens": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "createdAt": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Anmeldung aktivieren/deaktivieren (Passwort muss gesetzt sein)",
        "operationId": "putAuth",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["enabled"],
                "properties": { "enabled": { "type": "boolean" } }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Anmelden — setzt ein Session-Cookie (30 Tage)",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": { "password": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/auth/password": {
      "post": {
        "summary": "Passwort setzen oder ändern (invalidiert andere Sitzungen)",
        "operationId": "changePassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["new"],
                "properties": {
                  "current": { "type": "string" },
                  "new": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/auth/tokens": {
      "post": {
        "summary": "API-Token erzeugen (Klartext nur in dieser Antwort)",
        "description": "Verwendung: Header \"Authorization: Bearer <token>\". Gespeichert wird nur der SHA-256.",
        "operationId": "createToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": { "name": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Erzeugt",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "token": { "type": "string" } } }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/auth/tokens/{name}": {
      "delete": {
        "summary": "API-Token widerrufen",
        "operationId": "deleteToken",
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Ok" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/restore": {
      "post": {
        "summary": "Konfiguration aus einer Sicherung wiederherstellen",
        "description": "Ältere Sicherungen werden automatisch migriert und validiert; eine laufende Bewässerung stoppt.",
        "operationId": "postRestore",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": {
            "description": "Wiederhergestellt",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "restartRequired": { "type": "boolean" },
                    "outputError": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/system/run": {
      "put": {
        "summary": "Scheduler global an/aus",
        "operationId": "setRunSchedules",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["enabled"],
                "properties": { "enabled": { "type": "boolean" } }
              }
            }
          }
        },
        "responses": { "200": { "$ref": "#/components/responses/Ok" } }
      }
    },
    "/api/settings": {
      "get": {
        "summary": "Einstellungen lesen",
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "Einstellungen",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Settings" } } }
          }
        }
      },
      "put": {
        "summary": "Einstellungen ersetzen (vollständiges Objekt)",
        "operationId": "updateSettings",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Settings" } } }
        },
        "responses": {
          "200": {
            "description": "Gespeichert",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "restartRequired": { "type": "boolean", "description": "true, wenn der Web-Port geändert wurde" },
                    "outputError": { "type": "string", "description": "gesetzt, wenn das Hardware-Backend nicht initialisiert werden konnte" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "id": {
        "name": "id", "in": "path", "required": true,
        "schema": { "type": "integer", "minimum": 0 },
        "description": "0-basierter Index"
      }
    },
    "responses": {
      "Ok": {
        "description": "Erfolg",
        "content": {
          "application/json": {
            "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } }
          }
        }
      },
      "Error": {
        "description": "Fehler",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["error"],
              "properties": { "error": { "type": "string" } }
            }
          }
        }
      }
    },
    "schemas": {
...
      "ZoneInput": {
        "type": "object",
        "required": ["name", "enabled", "pump"],
        "properties": {
          "name": { "type": "string", "maxLength": 50 },
          "enabled": { "type": "boolean" },
          "pump": { "type": "boolean", "description": "Pumpe/Masterventil (Ausgang 0) mitschalten" }
        }
      },
      "Schedule": {
        "allOf": [
          { "$ref": "#/components/schemas/ScheduleInput" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "nextRun": {
                "oneOf": [{ "$ref": "#/components/schemas/NextRun" }, { "type": "null" }]
              }
            }
          }
        ]
      },
...
      "NextRun": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date" },
          "inDays": { "type": "integer" },
          "times": { "type": "array", "items": { "type": "integer" } }
        }
      },
...
    }
  }
}